fixed create class bugs

+++ server/controller/Department/deleteDept.js
@@ -8,21 +8,29 @@ module.exports = async (req, res) => {
       {
         _id: userId,
       },
-      { email: 1, _id: 0 }
+      { email: 1, _id: 0, classKeys: 1 }
     );
 
     if (dept) {
-      await accountModel.deleteOne({
-        email: dept.email,
-        type: "Department",
-      });
-      await deptModel.deleteOne({
-        _id: userId,
-      });
-      res.status(200).json({
-        status: "success",
-        message: "Department deleted",
-      });
+      if (dept.classKeys.length) {
+        res.status(400).json({
+          status: "failed",
+          message:
+            "Some classes are assigned to the Dept. Ensure first to delete the class",
+        });
+      } else {
+        await accountModel.deleteOne({
+          email: dept.email,
+          type: "Department",
+        });
+        await deptModel.deleteOne({
+          _id: userId,
+        });
+        res.status(200).json({
+          status: "success",
+          message: "Department deleted",
+        });
+      }
     } else {
       res.status(400).json({
         status: "failed",
